Move tweaked report sharding code to a separate file
