+++ integration_tests/tests/integration/simulation/mod.rs
@@ -71,14 +71,30 @@
 //! - Randomness used by `tokio::select!`.
 //! - Application-level concurrency bugs.
 
+use std::time::Duration;
+
+use backoff::ExponentialBackoff;
 use janus_messages::Time;
 
 const START_TIME: Time = Time::from_seconds_since_epoch(1_700_000_000);
 
 mod arbitrary;
+mod bad_client;
 mod model;
 mod proxy;
 mod quicktest;
 mod reproduction;
 mod run;
 mod setup;
+
+/// Aggressive exponential backoff parameters for this local-only test. Due to fault injection
+/// operations, we will often be hitting `max_elapsed_time`, so this value needs to be very low.
+pub(super) fn http_request_exponential_backoff() -> ExponentialBackoff {
+    ExponentialBackoff {
+        initial_interval: Duration::from_millis(10),
+        max_interval: Duration::from_millis(50),
+        multiplier: 2.0,
+        max_elapsed_time: Some(Duration::from_millis(250)),
+        ..Default::default()
+    }
+}
