+++ integration_tests/tests/integration/simulation/run.rs
@@ -6,11 +6,8 @@ use std::{
     time::Duration as StdDuration,
 };
 
-use backoff::ExponentialBackoff;
 use derivative::Derivative;
-use divviup_client::{Decode, Encode};
 use futures::future::join_all;
-use http::header::CONTENT_TYPE;
 use janus_aggregator::aggregator;
 use janus_aggregator_core::{
     datastore::models::AggregatorRole,
@@ -19,30 +16,23 @@ use janus_aggregator_core::{
 };
 use janus_collector::{Collection, CollectionJob, PollResult};
 use janus_core::{
-    hpke::{self, HpkeApplicationInfo, Label},
-    http::HttpErrorResponse,
-    retries::retry_http_request,
     test_util::runtime::TestRuntimeManager,
-    time::{Clock, MockClock, TimeExt},
+    time::{Clock, MockClock},
     vdaf::{vdaf_dp_strategies, VdafInstance},
 };
 use janus_messages::{
     query_type::{FixedSize, TimeInterval},
-    CollectionJobId, Duration, FixedSizeQuery, HpkeConfig, HpkeConfigList, InputShareAad,
-    PlaintextInputShare, Report, ReportId, ReportMetadata, Role, Time,
+    CollectionJobId, Duration, FixedSizeQuery, Time,
 };
 use opentelemetry::metrics::Meter;
-use prio::vdaf::{
-    prio3::{optimal_chunk_length, Prio3, Prio3Histogram},
-    Client as _,
-};
+use prio::vdaf::prio3::{optimal_chunk_length, Prio3, Prio3Histogram};
 use quickcheck::TestResult;
 use tokio::time::timeout;
 use tracing::{debug, error, info, info_span, warn, Instrument};
 use trillium_tokio::Stopper;
-use url::Url;
 
 use crate::simulation::{
+    bad_client::upload_replay_report,
     model::{Input, Op, Query},
     setup::Components,
     START_TIME,
@@ -540,97 +530,6 @@ impl State {
     }
 }
 
-/// Shard and upload a report, but with a fixed ReportId.
-async fn upload_replay_report(
-    measurement: usize,
-    task: &Task,
-    vdaf: &Prio3Histogram,
-    report_time: &Time,
-    http_client: &reqwest::Client,
-) -> Result<(), janus_client::Error> {
-    // This encodes to "replayreplayreplayrepl".
-    let report_id = ReportId::from([
-        173, 234, 101, 107, 42, 222, 166, 86, 178, 173, 234, 101, 107, 42, 222, 166,
-    ]);
-    let task_id = *task.id();
-    let (public_share, input_shares) = vdaf.shard(&measurement, report_id.as_ref())?;
-    let rounded_time = report_time
-        .to_batch_interval_start(task.time_precision())
-        .unwrap();
-    let report_metadata = ReportMetadata::new(report_id, rounded_time);
-    let encoded_public_share = public_share.get_encoded().unwrap();
-
-    let leader_hpke_config =
-        aggregator_hpke_config(task.leader_aggregator_endpoint(), http_client).await?;
-    let helper_hpke_config =
-        aggregator_hpke_config(task.helper_aggregator_endpoint(), http_client).await?;
-
-    let aad = InputShareAad::new(
-        task_id,
-        report_metadata.clone(),
-        encoded_public_share.clone(),
-    )
-    .get_encoded()?;
-    let leader_encrypted_input_share = hpke::seal(
-        &leader_hpke_config,
-        &HpkeApplicationInfo::new(&Label::InputShare, &Role::Client, &Role::Leader),
-        &PlaintextInputShare::new(Vec::new(), input_shares[0].get_encoded()?).get_encoded()?,
-        &aad,
-    )?;
-    let helper_encrypted_input_share = hpke::seal(
-        &helper_hpke_config,
-        &HpkeApplicationInfo::new(&Label::InputShare, &Role::Client, &Role::Helper),
-        &PlaintextInputShare::new(Vec::new(), input_shares[1].get_encoded()?).get_encoded()?,
-        &aad,
-    )?;
-
-    let report = Report::new(
-        report_metadata,
-        encoded_public_share,
-        leader_encrypted_input_share,
-        helper_encrypted_input_share,
-    );
-
-    let url = task
-        .leader_aggregator_endpoint()
-        .join(&format!("tasks/{task_id}/reports"))
-        .unwrap();
-    retry_http_request(http_request_exponential_backoff(), || async {
-        http_client
-            .put(url.clone())
-            .header(CONTENT_TYPE, Report::MEDIA_TYPE)
-            .body(report.get_encoded().unwrap())
-            .send()
-            .await
-    })
-    .await?;
-
-    Ok(())
-}
-
-async fn aggregator_hpke_config(
-    endpoint: &Url,
-    http_client: &reqwest::Client,
-) -> Result<HpkeConfig, janus_client::Error> {
-    let response = retry_http_request(http_request_exponential_backoff(), || async {
-        http_client
-            .get(endpoint.join("hpke_config").unwrap())
-            .send()
-            .await
-    })
-    .await?;
-    let status = response.status();
-    if !status.is_success() {
-        return Err(janus_client::Error::Http(Box::new(
-            HttpErrorResponse::from(status),
-        )));
-    }
-
-    let list = HpkeConfigList::get_decoded(response.body())?;
-
-    Ok(list.hpke_configs()[0].clone())
-}
-
 fn check_aggregate_results_valid<Q: janus_messages::query_type::QueryType>(
     map: &HashMap<CollectionJobId, Collection<Vec<u128>, Q>>,
     state: &State,
@@ -654,16 +553,4 @@ fn check_aggregate_results_valid<Q: janus_messages::query_type::QueryType>(
         }
     }
     true
-}
-
-/// Aggressive exponential backoff parameters for this local-only test. Due to fault injection
-/// operations, we will often be hitting `max_elapsed_time`, so this value needs to be very low.
-pub(super) fn http_request_exponential_backoff() -> ExponentialBackoff {
-    ExponentialBackoff {
-        initial_interval: StdDuration::from_millis(10),
-        max_interval: StdDuration::from_millis(50),
-        multiplier: 2.0,
-        max_elapsed_time: Some(StdDuration::from_millis(250)),
-        ..Default::default()
-    }
 }
